+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 - Add containers prune button to remove all stopped containers
 - Add a way to follow container logs, the logs are now paginated
 - Make container and image ids open their respective page with details
+- Add a button to clear images build cache